Rename old int-based frac into counting permission.
