+++ vstd_extra/src/resource/ghost_resource/mod.rs
@@ -1,5 +1,5 @@
 //! Ghost resources available for use in proofs.
 pub mod csum;
+pub mod count;
 pub mod excl;
-pub mod frac;
 pub mod tokens;
