+++ vstd_extra/src/resource/ghost_resource/tokens.rs
@@ -1,551 +1,26 @@
 use std::borrow::Borrow;
 
-//！ A wrapper around `vstd::tokens::FracGhost` that stores and dispatches fractional access.
 use crate::sum::*;
 use vstd::map::*;
 use vstd::modes::tracked_swap;
 use vstd::prelude::*;
-use vstd::resource::algebra::ResourceAlgebra;
-use vstd::resource::pcm::{Resource, PCM};
-use vstd::resource::storage_protocol::{Protocol, StorageResource};
 use vstd::resource::Loc;
+use crate::resource::ghost_resource::count::*;
 
 verus! {
 
 broadcast use {vstd::map::group_map_axioms, vstd::set::group_set_axioms};
 
-// Integer-based fractional ghost tokens kept for code that predates
-// vstd::resource::frac's real-valued fractions.
-ghost enum FractionalCarrier<T, const TOTAL: u64> {
-    Value { v: T, n: int },
-    Empty,
-    Invalid,
-}
-
-impl<T, const TOTAL: u64> FractionalCarrier<T, TOTAL> {
-    spec fn new(v: T) -> Self {
-        FractionalCarrier::Value { v, n: TOTAL as int }
-    }
-}
-
-impl<T, const TOTAL: u64> ResourceAlgebra for FractionalCarrier<T, TOTAL> {
-    closed spec fn valid(self) -> bool {
-        match self {
-            FractionalCarrier::Invalid => false,
-            FractionalCarrier::Empty => true,
-            FractionalCarrier::Value { v: _, n } => 0 < n <= TOTAL,
-        }
-    }
-
-    closed spec fn op(a: Self, b: Self) -> Self {
-        match a {
-            FractionalCarrier::Invalid => FractionalCarrier::Invalid,
-            FractionalCarrier::Empty => b,
-            FractionalCarrier::Value { v: sv, n: sn } => match b {
-                FractionalCarrier::Invalid => FractionalCarrier::Invalid,
-                FractionalCarrier::Empty => a,
-                FractionalCarrier::Value { v: ov, n: on } => {
-                    if sv != ov {
-                        FractionalCarrier::Invalid
-                    } else if sn <= 0 || on <= 0 {
-                        FractionalCarrier::Invalid
-                    } else {
-                        FractionalCarrier::Value { v: sv, n: sn + on }
-                    }
-                },
-            },
-        }
-    }
-
-    proof fn valid_op(a: Self, b: Self) {
-    }
-
-    proof fn commutative(a: Self, b: Self) {
-    }
-
-    proof fn associative(a: Self, b: Self, c: Self) {
-    }
-}
-
-impl<T, const TOTAL: u64> PCM for FractionalCarrier<T, TOTAL> {
-    closed spec fn unit() -> Self {
-        FractionalCarrier::Empty
-    }
-
-    proof fn op_unit(self) {
-    }
-
-    proof fn unit_valid() {
-    }
-}
-
-pub tracked struct FracGhost<T, const TOTAL: u64 = 2> {
-    r: Resource<FractionalCarrier<T, TOTAL>>,
-}
-
-impl<T, const TOTAL: u64> FracGhost<T, TOTAL> {
-    #[verifier::type_invariant]
-    spec fn inv(self) -> bool {
-        self.r.value() is Value
-    }
-
-    pub closed spec fn id(self) -> Loc {
-        self.r.loc()
-    }
-
-    pub closed spec fn view(self) -> T {
-        self.r.value()->v
-    }
-
-    pub closed spec fn frac(self) -> int {
-        self.r.value()->n
-    }
-
-    pub open spec fn valid(self, id: Loc, frac: int) -> bool {
-        &&& self.id() == id
-        &&& self.frac() == frac
-    }
-
-    pub proof fn new(v: T) -> (tracked result: Self)
-        requires
-            TOTAL > 0,
-        ensures
-            result.frac() == TOTAL as int,
-            result@ == v,
-    {
-        let f = FractionalCarrier::<T, TOTAL>::new(v);
-        let tracked r = Resource::alloc(f);
-        Self { r }
-    }
-
-    pub proof fn agree(tracked self: &Self, tracked other: &Self)
-        requires
-            self.id() == other.id(),
-        ensures
-            self@ == other@,
-    {
-        use_type_invariant(self);
-        use_type_invariant(other);
-        let tracked joined = self.r.join_shared(&other.r);
-        joined.validate()
-    }
-
-    pub proof fn take(tracked &mut self) -> (tracked result: Self)
-        ensures
-            result == *old(self),
-    {
-        self.bounded();
-        let tracked mut mself = Self::dummy();
-        tracked_swap(self, &mut mself);
-        mself
-    }
-
-    pub proof fn split(tracked &mut self, n: int) -> (tracked result: Self)
-        requires
-            0 < n < old(self).frac(),
-        ensures
-            result.id() == final(self).id(),
-            final(self).id() == old(self).id(),
-            final(self)@ == old(self)@,
-            result@ == old(self)@,
-            final(self).frac() + result.frac() == old(self).frac(),
-            result.frac() == n,
-    {
-        self.bounded();
-        let tracked mut mself = Self::dummy();
-        tracked_swap(self, &mut mself);
-        use_type_invariant(&mself);
-        let tracked (r1, r2) = mself.r.split(
-            FractionalCarrier::Value { v: mself.r.value()->v, n: mself.r.value()->n - n },
-            FractionalCarrier::Value { v: mself.r.value()->v, n },
-        );
-        self.r = r1;
-        Self { r: r2 }
-    }
-
-    pub proof fn combine(tracked &mut self, tracked other: Self)
-        requires
-            old(self).id() == other.id(),
-        ensures
-            final(self).id() == old(self).id(),
-            final(self)@ == old(self)@,
-            final(self)@ == other@,
-            final(self).frac() == old(self).frac() + other.frac(),
-    {
-        self.bounded();
-        let tracked mut mself = Self::dummy();
-        tracked_swap(self, &mut mself);
-        use_type_invariant(&mself);
-        use_type_invariant(&other);
-        let tracked mut r = mself.r;
-        r.validate_2(&other.r);
-        *self = Self { r: r.join(other.r) };
-    }
-
-    pub proof fn update(tracked &mut self, v: T)
-        requires
-            old(self).frac() == TOTAL,
-        ensures
-            final(self).id() == old(self).id(),
-            final(self)@ == v,
-            final(self).frac() == old(self).frac(),
-    {
-        self.bounded();
-        let tracked mut mself = Self::dummy();
-        tracked_swap(self, &mut mself);
-        use_type_invariant(&mself);
-        let tracked r = mself.r;
-        let f = FractionalCarrier::<T, TOTAL>::Value { v, n: TOTAL as int };
-        *self = Self { r: r.update(f) };
-    }
-
-    pub proof fn update_with(tracked &mut self, tracked other: &mut Self, v: T)
-        requires
-            old(self).id() == old(other).id(),
-            old(self).frac() + old(other).frac() == TOTAL,
-        ensures
-            final(self).id() == old(self).id(),
-            final(other).id() == old(other).id(),
-            final(self).frac() == old(self).frac(),
-            final(other).frac() == old(other).frac(),
-            old(self)@ == old(other)@,
-            final(self)@ == v,
-            final(other)@ == v,
-    {
-        let ghost other_frac = other.frac();
-        other.bounded();
-        let tracked mut xother = Self::dummy();
-        tracked_swap(other, &mut xother);
-        self.bounded();
-        self.combine(xother);
-        self.update(v);
-        let tracked mut xother = self.split(other_frac);
-        tracked_swap(other, &mut xother);
-    }
-
-    pub proof fn bounded(tracked &self)
-        ensures
-            0 < self.frac() <= TOTAL,
-    {
-        use_type_invariant(self);
-        self.r.validate()
-    }
-
-    pub proof fn dummy() -> (tracked result: Self)
-        requires
-            TOTAL > 0,
-    {
-        Self::new(arbitrary())
-    }
-}
-
-ghost enum FractionalCarrierOpt<T, const TOTAL: u64> {
-    Value { v: Option<T>, n: int },
-    Empty,
-    Invalid,
-}
-
-impl<T, const TOTAL: u64> Protocol<(), T> for FractionalCarrierOpt<T, TOTAL> {
-    closed spec fn op(self, other: Self) -> Self {
-        match self {
-            FractionalCarrierOpt::Invalid => FractionalCarrierOpt::Invalid,
-            FractionalCarrierOpt::Empty => other,
-            FractionalCarrierOpt::Value { v: sv, n: sn } => match other {
-                FractionalCarrierOpt::Invalid => FractionalCarrierOpt::Invalid,
-                FractionalCarrierOpt::Empty => self,
-                FractionalCarrierOpt::Value { v: ov, n: on } => {
-                    if sv != ov {
-                        FractionalCarrierOpt::Invalid
-                    } else if sn <= 0 || on <= 0 {
-                        FractionalCarrierOpt::Invalid
-                    } else {
-                        FractionalCarrierOpt::Value { v: sv, n: sn + on }
-                    }
-                },
-            },
-        }
-    }
-
-    closed spec fn rel(self, s: Map<(), T>) -> bool {
-        match self {
-            FractionalCarrierOpt::Value { v, n } => {
-                (match v {
-                    Some(v0) => s.dom().contains(()) && s[()] == v0,
-                    None => s =~= map![],
-                }) && n == TOTAL && n != 0
-            },
-            FractionalCarrierOpt::Empty => false,
-            FractionalCarrierOpt::Invalid => false,
-        }
-    }
-
-    closed spec fn unit() -> Self {
-        FractionalCarrierOpt::Empty
-    }
-
-    proof fn commutative(a: Self, b: Self) {
-    }
-
-    proof fn associative(a: Self, b: Self, c: Self) {
-    }
-
-    proof fn op_unit(a: Self) {
-    }
-}
-
-pub tracked struct Frac<T, const TOTAL: u64 = 2> {
-    r: StorageResource<(), T, FractionalCarrierOpt<T, TOTAL>>,
-}
-
-pub tracked struct Empty<T, const TOTAL: u64 = 2> {
-    r: StorageResource<(), T, FractionalCarrierOpt<T, TOTAL>>,
-}
-
-impl<T, const TOTAL: u64> Frac<T, TOTAL> {
-    #[verifier::type_invariant]
-    spec fn inv(self) -> bool {
-        self.r.value() matches FractionalCarrierOpt::Value { v: Some(_), .. }
-    }
-
-    pub closed spec fn id(self) -> Loc {
-        self.r.loc()
-    }
-
-    pub closed spec fn resource(self) -> T {
-        self.r.value()->v.unwrap()
-    }
-
-    pub closed spec fn frac(self) -> int {
-        self.r.value()->n
-    }
-
-    pub open spec fn valid(self, id: Loc, frac: int) -> bool {
-        &&& self.id() == id
-        &&& self.frac() == frac
-    }
-
-    pub proof fn new(tracked v: T) -> (tracked result: Self)
-        requires
-            TOTAL > 0,
-        ensures
-            result.frac() == TOTAL as int,
-            result.resource() == v,
-    {
-        let f = FractionalCarrierOpt::<T, TOTAL>::Value { v: Some(v), n: TOTAL as int };
-        let tracked mut m = Map::<(), T>::tracked_empty();
-        m.tracked_insert((), v);
-        let tracked r = StorageResource::alloc(f, m);
-        Self { r }
-    }
-
-    pub proof fn agree(tracked self: &Self, tracked other: &Self)
-        requires
-            self.id() == other.id(),
-        ensures
-            self.resource() == other.resource(),
-    {
-        use_type_invariant(self);
-        use_type_invariant(other);
-        let tracked joined = self.r.join_shared(&other.r);
-        joined.validate();
-    }
-
-    pub proof fn split(tracked &mut self, n: int) -> (tracked result: Self)
-        requires
-            0 < n < old(self).frac(),
-        ensures
-            result.id() == final(self).id(),
-            final(self).id() == old(self).id(),
-            final(self).resource() == old(self).resource(),
-            result.resource() == old(self).resource(),
-            final(self).frac() + result.frac() == old(self).frac(),
-            result.frac() == n,
-    {
-        use_type_invariant(&*self);
-        Self::split_helper(&mut self.r, n)
-    }
-
-    proof fn split_helper(
-        tracked r: &mut StorageResource<(), T, FractionalCarrierOpt<T, TOTAL>>,
-        n: int,
-    ) -> (tracked result: Self)
-        requires
-            0 < n < old(r).value()->n,
-            old(r).value() matches FractionalCarrierOpt::Value { v: Some(_), .. },
-        ensures
-            result.id() == final(r).loc(),
-            final(r).loc() == old(r).loc(),
-            final(r).value()->v.unwrap() == old(r).value()->v.unwrap(),
-            result.resource() == old(r).value()->v.unwrap(),
-            final(r).value()->n + result.frac() == old(r).value()->n,
-            result.frac() == n,
-            final(r).value() matches FractionalCarrierOpt::Value { v: Some(_), .. },
-    {
-        r.validate();
-        let tracked mut r1 = StorageResource::alloc(
-            FractionalCarrierOpt::Value { v: None, n: TOTAL as int },
-            Map::tracked_empty(),
-        );
-        tracked_swap(r, &mut r1);
-        let tracked (r1, r2) = r1.split(
-            FractionalCarrierOpt::Value { v: r1.value()->v, n: r1.value()->n - n },
-            FractionalCarrierOpt::Value { v: r1.value()->v, n },
-        );
-        *r = r1;
-        Self { r: r2 }
-    }
-
-    pub proof fn combine(tracked &mut self, tracked other: Self)
-        requires
-            old(self).id() == other.id(),
-        ensures
-            final(self).id() == old(self).id(),
-            final(self).resource() == old(self).resource(),
-            final(self).resource() == other.resource(),
-            final(self).frac() == old(self).frac() + other.frac(),
-    {
-        use_type_invariant(&*self);
-        Self::combine_helper(&mut self.r, other)
-    }
-
-    proof fn combine_helper(
-        tracked r: &mut StorageResource<(), T, FractionalCarrierOpt<T, TOTAL>>,
-        tracked other: Self,
-    )
-        requires
-            old(r).loc() == other.id(),
-            old(r).value() matches FractionalCarrierOpt::Value { v: Some(_), .. },
-        ensures
-            final(r).loc() == old(r).loc(),
-            final(r).value()->v.unwrap() == old(r).value()->v.unwrap(),
-            final(r).value()->v.unwrap() == other.resource(),
-            final(r).value()->n == old(r).value()->n + other.frac(),
-            final(r).value() matches FractionalCarrierOpt::Value { v: Some(_), .. },
-    {
-        r.validate();
-        use_type_invariant(&other);
-        let tracked mut r1 = StorageResource::alloc(
-            FractionalCarrierOpt::Value { v: None, n: TOTAL as int },
-            Map::tracked_empty(),
-        );
-        tracked_swap(r, &mut r1);
-        r1.validate_with_shared(&other.r);
-        *r = StorageResource::join(r1, other.r);
-    }
-
-    pub proof fn bounded(tracked &self)
-        ensures
-            0 < self.frac() <= TOTAL,
-    {
-        use_type_invariant(self);
-        let (x, _) = self.r.validate();
-    }
-
-    pub proof fn borrow(tracked &self) -> (tracked ret: &T)
-        ensures
-            ret == self.resource(),
-    {
-        use_type_invariant(self);
-        StorageResource::guard(&self.r, map![() => self.resource()]).tracked_borrow(())
-    }
-
-    pub proof fn take_resource(tracked self) -> (tracked pair: (T, Empty<T, TOTAL>))
-        requires
-            self.frac() == TOTAL,
-        ensures
-            pair.0 == self.resource(),
-            pair.1.id() == self.id(),
-    {
-        use_type_invariant(&self);
-        self.r.validate();
-        let p1 = self.r.value();
-        let p2 = FractionalCarrierOpt::Value { v: None, n: TOTAL as int };
-        let b2 = map![() => self.resource()];
-        assert forall|q: FractionalCarrierOpt<T, TOTAL>, t1: Map<(), T>|
-            #![all_triggers]
-            FractionalCarrierOpt::rel(FractionalCarrierOpt::op(p1, q), t1) implies exists|
-            t2: Map<(), T>,
-        |
-            #![all_triggers]
-            FractionalCarrierOpt::rel(FractionalCarrierOpt::op(p2, q), t2) && t2.dom().disjoint(
-                b2.dom(),
-            ) && t1 =~= t2.union_prefer_right(b2) by {
-            let t2 = map![];
-            assert(FractionalCarrierOpt::rel(FractionalCarrierOpt::op(p2, q), t2));
-            assert(t2.dom().disjoint(b2.dom()));
-            assert(t1 =~= t2.union_prefer_right(b2));
-        }
-        let tracked Self { r } = self;
-        let tracked (new_r, mut m) = r.withdraw(p2, b2);
-        let tracked emp = Empty { r: new_r };
-        let tracked resource = m.tracked_remove(());
-        (resource, emp)
-    }
-}
-
-impl<T, const TOTAL: u64> Empty<T, TOTAL> {
-    #[verifier::type_invariant]
-    spec fn inv(self) -> bool {
-        &&& self.r.value() matches FractionalCarrierOpt::Value { v: None, n }
-        &&& n == TOTAL
-    }
-
-    pub closed spec fn id(self) -> Loc {
-        self.r.loc()
-    }
-
-    pub proof fn new(tracked v: T) -> (tracked result: Self)
-        requires
-            TOTAL > 0,
-    {
-        let f = FractionalCarrierOpt::<T, TOTAL>::Value { v: None, n: TOTAL as int };
-        let tracked mut m = Map::<(), T>::tracked_empty();
-        let tracked r = StorageResource::alloc(f, m);
-        Self { r }
-    }
-
-    pub proof fn put_resource(tracked self, tracked resource: T) -> (tracked frac: Frac<T, TOTAL>)
-        ensures
-            frac.id() == self.id(),
-            frac.resource() == resource,
-            frac.frac() == TOTAL,
-    {
-        use_type_invariant(&self);
-        self.r.validate();
-        let p1 = self.r.value();
-        let b1 = map![() => resource];
-        let p2 = FractionalCarrierOpt::Value { v: Some(resource), n: TOTAL as int };
-        assert forall|q: FractionalCarrierOpt<T, TOTAL>, t1: Map<(), T>|
-            #![all_triggers]
-            FractionalCarrierOpt::rel(FractionalCarrierOpt::op(p1, q), t1) implies exists|
-            t2: Map<(), T>,
-        |
-            #![all_triggers]
-            FractionalCarrierOpt::rel(FractionalCarrierOpt::op(p2, q), t2) && t1.dom().disjoint(
-                b1.dom(),
-            ) && t1.union_prefer_right(b1) =~= t2 by {
-            let t2 = map![() => resource];
-            assert(FractionalCarrierOpt::rel(FractionalCarrierOpt::op(p2, q), t2)
-                && t1.dom().disjoint(b1.dom()) && t1.union_prefer_right(b1) =~= t2);
-        }
-        let tracked mut m = Map::tracked_empty();
-        m.tracked_insert((), resource);
-        let tracked Self { r } = self;
-        let tracked new_r = r.deposit(m, p2);
-        Frac { r: new_r }
-    }
-}
-
-/// A struct that stores and dispatches `FracGhost<T>`.
-/// Unlike `FracGhost`, it provides an `empty` state.
+/// A struct that stores and dispatches `CountGhost<T>`.
+/// Unlike `CountGhost`, it provides an `empty` state.
 /// It also remembers the value even it is empty.
-pub tracked struct FracGhostResource<T, const TOTAL: u64> {
-    tracked r: Option<FracGhost<T, TOTAL>>,
+pub tracked struct CountGhostResource<T, const TOTAL: u64> {
+    tracked r: Option<CountGhost<T, TOTAL>>,
     ghost snapshot: T,
     ghost id: Loc,
 }
 
-impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
+impl<T, const TOTAL: u64> CountGhostResource<T, TOTAL> {
     #[verifier::type_invariant]
     closed spec fn type_inv(self) -> bool {
         &&& TOTAL > 0
@@ -562,32 +37,32 @@ impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
         &&& 0 <= self.frac() <= TOTAL
     }
 
-    /// Whether this `FracGhostResource` is empty, i.e., has no fraction.
+    /// Whether this `CountGhostResource` is empty, i.e., has no fraction.
     pub open spec fn is_empty(self) -> bool {
         self.frac() == 0
     }
 
-    /// Whether the fraction stored in this `FracGhostResource` is less than `TOTAL`.
+    /// Whether the fraction stored in this `CountGhostResource` is less than `TOTAL`.
     pub open spec fn not_empty(self) -> bool {
         !self.is_empty()
     }
 
-    /// Whether this `FracGhostResource` has the full fraction, i.e., `TOTAL`.
+    /// Whether this `CountGhostResource` has the full fraction, i.e., `TOTAL`.
     pub open spec fn is_full(self) -> bool {
         self.frac() == TOTAL
     }
 
-    /// Returns the `FracGhost<T,TOTAL>` stored in this `FracGhostResource`.
-    pub closed spec fn storage(self) -> FracGhost<T, TOTAL> {
+    /// Returns the `CountGhost<T,TOTAL>` stored in this `CountGhostResource`.
+    pub closed spec fn storage(self) -> CountGhost<T, TOTAL> {
         self.r->Some_0
     }
 
-    /// Returns the value of type `T` stored in this `FracGhostResource`.
+    /// Returns the value of type `T` stored in this `CountGhostResource`.
     pub closed spec fn view(self) -> T {
         self.snapshot
     }
 
-    /// The fractions stored in this `FracGhostResource`.
+    /// The fractions stored in this `CountGhostResource`.
     pub closed spec fn frac(self) -> int {
         if self.r is None {
             0int
@@ -601,7 +76,7 @@ impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
         self.id
     }
 
-    /// Create an arbitrary `FracGhostResource`. Useful as a placeholder.
+    /// Create an arbitrary `CountGhostResource`. Useful as a placeholder.
     pub proof fn arbitrary() -> (tracked res: Self)
         requires
             TOTAL > 0,
@@ -609,7 +84,7 @@ impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
         Self { r: None, snapshot: arbitrary(), id: arbitrary() }
     }
 
-    /// Allocates a new `FracGhostResource` with the full fraction and the given value.
+    /// Allocates a new `CountGhostResource` with the full fraction and the given value.
     pub proof fn alloc(value: T) -> (tracked res: Self)
         requires
             TOTAL > 0,
@@ -619,12 +94,12 @@ impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
             res@ == value,
             res.wf(),
     {
-        let tracked r = FracGhost::new(value);
+        let tracked r = CountGhost::new(value);
         Self { r: Some(r), snapshot: value, id: r.id() }
     }
 
-    /// Splits a `FracGhost` with fraction 1.
-    pub proof fn split_one(tracked &mut self) -> (tracked res: FracGhost<T, TOTAL>)
+    /// Splits a `CountGhost` with fraction 1.
+    pub proof fn split_one(tracked &mut self) -> (tracked res: CountGhost<T, TOTAL>)
         requires
             old(self).not_empty(),
         ensures
@@ -649,8 +124,8 @@ impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
         }
     }
 
-    /// Splits a `FracGhost` with the given fraction.
-    pub proof fn split(tracked &mut self, n: int) -> (tracked res: FracGhost<T, TOTAL>)
+    /// Splits a `CountGhost` with the given fraction.
+    pub proof fn split(tracked &mut self, n: int) -> (tracked res: CountGhost<T, TOTAL>)
         requires
             1 <= n <= old(self).frac(),
         ensures
@@ -675,8 +150,8 @@ impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
         }
     }
 
-    /// Combines a `FracGhost`.
-    pub proof fn combine(tracked &mut self, tracked other: FracGhost<T, TOTAL>)
+    /// Combines a `CountGhost`.
+    pub proof fn combine(tracked &mut self, tracked other: CountGhost<T, TOTAL>)
         requires
             old(self).id() == other.id(),
             other@ == old(self)@,
@@ -701,7 +176,7 @@ impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
         }
     }
 
-    /// `FracGhostResource` satisfies the type invariant.
+    /// `CountGhostResource` satisfies the type invariant.
     pub proof fn validate(tracked &self)
         ensures
             self.wf(),
@@ -725,7 +200,7 @@ impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
         }
     }
 
-    /// Updates the value stored in this `FracGhostResource`.
+    /// Updates the value stored in this `CountGhostResource`.
     /// The fraction must be full before the update.
     pub proof fn update(tracked &mut self, value: T)
         requires
@@ -744,18 +219,18 @@ impl<T, const TOTAL: u64> FracGhostResource<T, TOTAL> {
     }
 }
 
-pub type TokenResource<const TOTAL: u64> = FracGhostResource<(), TOTAL>;
+pub type TokenResource<const TOTAL: u64> = CountGhostResource<(), TOTAL>;
 
-pub type Token<const TOTAL: u64> = FracGhost<(), TOTAL>;
+pub type Token<const TOTAL: u64> = CountGhost<(), TOTAL>;
 
 /// A struct that stores and dispatches `Frac<T>`.
 /// Unlike `Frac`, it provides an `empty` state.
-pub tracked struct FracResource<T, const TOTAL: u64> {
-    tracked r: Option<Frac<T, TOTAL>>,
+pub tracked struct CountResource<T, const TOTAL: u64> {
+    tracked r: Option<Count<T, TOTAL>>,
     ghost id: Loc,
 }
 
-impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
+impl<T, const TOTAL: u64> CountResource<T, TOTAL> {
     #[verifier::type_invariant]
     pub closed spec fn type_inv(self) -> bool {
         &&& TOTAL > 0
@@ -766,8 +241,8 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         }
     }
 
-    /// Returns the `Frac<T,TOTAL>` stored in this `FracResource`.
-    pub closed spec fn storage(self) -> Frac<T, TOTAL> {
+    /// Returns the `Count<T,TOTAL>` stored in this `CountResource`.
+    pub closed spec fn storage(self) -> Count<T, TOTAL> {
         self.r->Some_0
     }
 
@@ -778,33 +253,33 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         &&& self.type_inv()
     }
 
-    /// Whether this `FracResource` is empty, i.e., has no fraction.
+    /// Whether this `CountResource` is empty, i.e., has no fraction.
     pub open spec fn is_empty(self) -> bool {
         self.frac() == 0
     }
 
-    /// Whether the fraction stored in this `FracResource` is less than `TOTAL`.
+    /// Whether the fraction stored in this `CountResource` is less than `TOTAL`.
     pub open spec fn not_empty(self) -> bool {
         !self.is_empty()
     }
 
-    /// Whether this `FracResource` has the full fraction, i.e., `TOTAL`.
+    /// Whether this `CountResource` has the full fraction, i.e., `TOTAL`.
     pub open spec fn is_full(self) -> bool {
         self.frac() == TOTAL
     }
 
-    /// Returns the value of type `T` stored in this `FracResource`.
+    /// Returns the value of type `T` stored in this `CountResource`.
     pub closed spec fn resource(self) -> T {
         self.storage().resource()
     }
 
-    /// Returns the value of type `T` stored in this `FracResource`. It is an alias of `Self::resource`.
+    /// Returns the value of type `T` stored in this `CountResource`. It is an alias of `Self::resource`.
     #[verifier::inline]
     pub open spec fn view(self) -> T {
         self.resource()
     }
 
-    /// The fractions stored in this `FracResource`.
+    /// The fractions stored in this `CountResource`.
     pub closed spec fn frac(self) -> int {
         if self.r is None {
             0int
@@ -818,7 +293,7 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         self.id
     }
 
-    /// Create an arbitrary `FracResource`. Useful as a placeholder.
+    /// Create an arbitrary `CountResource`. Useful as a placeholder.
     pub proof fn arbitrary() -> (tracked res: Self)
         requires
             TOTAL > 0,
@@ -826,7 +301,7 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         Self { r: None, id: arbitrary() }
     }
 
-    /// Allocates a new `FracResource` with the given tracked object.
+    /// Allocates a new `CountResource` with the given tracked object.
     pub proof fn alloc(tracked value: T) -> (tracked res: Self)
         requires
             TOTAL > 0,
@@ -836,12 +311,12 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
             res@ == value,
             res.wf(),
     {
-        let tracked r = Frac::new(value);
+        let tracked r = Count::new(value);
         Self { r: Some(r), id: r.id() }
     }
 
-    /// Allocates a new `FracResource` from an `Empty<T,TOTAL>` with the given tracked object.
-    pub proof fn alloc_from_empty(tracked empty: Empty<T, TOTAL>, tracked value: T) -> (tracked res:
+    /// Allocates a new `CountResource` from an `EmptyCount<T,TOTAL>` with the given tracked object.
+    pub proof fn alloc_from_empty(tracked empty: EmptyCount<T, TOTAL>, tracked value: T) -> (tracked res:
         Self)
         requires
             TOTAL > 0,
@@ -855,8 +330,8 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         Self { r: Some(r), id: r.id() }
     }
 
-    /// Splits a `Frac` with fraction 1.
-    pub proof fn split_one(tracked &mut self) -> (tracked res: Frac<T, TOTAL>)
+    /// Splits a `Count` with fraction 1.
+    pub proof fn split_one(tracked &mut self) -> (tracked res: Count<T, TOTAL>)
         requires
             old(self).not_empty(),
         ensures
@@ -881,8 +356,8 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         }
     }
 
-    /// Splits a `Frac` with the given fraction.
-    pub proof fn split(tracked &mut self, n: int) -> (tracked res: Frac<T, TOTAL>)
+    /// Splits a `Count` with the given fraction.
+    pub proof fn split(tracked &mut self, n: int) -> (tracked res: Count<T, TOTAL>)
         requires
             1 <= n <= old(self).frac(),
         ensures
@@ -907,8 +382,8 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         }
     }
 
-    /// Combines a `Frac`.
-    pub proof fn combine(tracked &mut self, tracked other: Frac<T, TOTAL>)
+    /// Combines a `Count`.
+    pub proof fn combine(tracked &mut self, tracked other: Count<T, TOTAL>)
         requires
             old(self).id() == other.id(),
         ensures
@@ -934,7 +409,7 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         }
     }
 
-    /// `FracResource` satisfies the type invariant.
+    /// `CountResource` satisfies the type invariant.
     pub proof fn validate(tracked &self)
         ensures
             self.wf(),
@@ -958,7 +433,7 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         }
     }
 
-    pub proof fn validate_with_frac(tracked &self, tracked frac: &Frac<T, TOTAL>)
+    pub proof fn validate_with_frac(tracked &self, tracked frac: &Count<T, TOTAL>)
         requires
             self.id() == frac.id(),
             self.frac() > 0,
@@ -970,7 +445,7 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
     }
 
     /// Consumes the token and takes out the resource.
-    pub proof fn take_resource(tracked self) -> (tracked (res, empty): (T, Empty<T, TOTAL>))
+    pub proof fn take_resource(tracked self) -> (tracked (res, empty): (T, EmptyCount<T, TOTAL>))
         requires
             self.is_full(),
         ensures
@@ -982,7 +457,7 @@ impl<T, const TOTAL: u64> FracResource<T, TOTAL> {
         r.take_resource()
     }
 
-    /// Updates the resource stored in this `FracResource` and retunrs the old resource if it exists.
+    /// Updates the resource stored in this `CountResource` and retunrs the old resource if it exists.
     /// The fraction must be full before the update.
     pub proof fn update(tracked &mut self, tracked value: T) -> (tracked res: T)
         requires
