+++ vstd_extra/src/resource/storage_protocol/mod.rs
@@ -3,5 +3,4 @@
 //! The generic storage-protocol API itself now comes from `vstd::resource::storage_protocol`.
 pub mod csum;
 pub mod excl;
-pub mod frac;
 pub mod hybrid_product;
